PMM-6705 Adding tests for Suspend Resume
