+++ pmm-app/codeceptConfigHelper.js
@@ -13,6 +13,7 @@ module.exports = {
     databaseChecksPage: './tests/pages/databaseChecksPage.js',
     dbaasAPI: './tests/DbaaS/api/dbaasAPI.js',
     dbaasPage: './tests/DbaaS/pages/dbaasPage.js',
+    dbaasActionsPage: './tests/DbaaS/pages/dbaasActionsPage.js',
     homePage: './tests/pages/homePage.js',
     inventoryAPI: './tests/pages/api/inventoryAPI.js',
     mysqlTableDetailsPage: './tests/pages/mysqlTableDetailsPage.js',
